test(js): sync mirrored test helpers with refactored production code
- Extract parseTypoLink/parseTypoLinkParts/encodeTypoLink from
  typo3image.js into a sibling module typolink-parser.js (mirrors the
  existing sanitize-src.js precedent). typo3image.js re-exports them
  for backwards compatibility. This was necessary because typo3image.js
  has top-level CKEditor imports that cannot resolve in the unit-test
  environment, so tests cannot import directly from it.
- typolink-parser.test.js: import parseTypoLink/parseTypoLinkParts/
  encodeTypoLink from the new sibling module instead of duplicating
  ~180 lines. This both eliminates the source of "duplication on new
  code" and makes future drift impossible.
- linked-image-upcast.test.js: update upcast-converter mirrored helpers
  to use optional chaining matching production refactor.
- editing-downcast.test.js: update downcast-converter mirrored helpers
  similarly.

Addresses Copilot review feedback on PR #813 that mirrored test copies
weren't keeping up with production changes, allowing the test suite to
stay green while production logic regresses.
