+++ Tests/JavaScript/tests/typolink-parser.test.js
@@ -14,173 +14,11 @@
  */
 
 import { describe, it, expect } from 'vitest';
-
-/**
- * Parse TypoLink string into parts using CSV-like logic.
- * Handles quoted strings with spaces and escaped characters.
- *
- * This is a copy of the production code for testing purposes.
- * @see Resources/Public/JavaScript/Plugins/typo3image.js
- *
- * @param {string} str - The TypoLink string
- * @return {string[]} Array of parsed parts
- */
-function parseTypoLinkParts(str) {
-    const parts = [];
-    let current = '';
-    let inQuotes = false;
-    let i = 0;
-
-    while (i < str.length) {
-        const char = str[i];
-
-        if (char === '\\' && i + 1 < str.length) {
-            // Escape sequence - include next character literally
-            current += str[i + 1];
-            i += 2;
-            continue;
-        }
-
-        if (char === '"') {
-            inQuotes = !inQuotes;
-            i++;
-            continue;
-        }
-
-        if (char === ' ' && !inQuotes) {
-            // Delimiter - save current part and start new one
-            if (current !== '' || parts.length > 0) {
-                parts.push(current);
-                current = '';
-            }
-            i++;
-            continue;
-        }
-
-        current += char;
-        i++;
-    }
-
-    // Don't forget the last part
-    if (current !== '' || parts.length > 0) {
-        parts.push(current);
-    }
-
-    return parts;
-}
-
-/**
- * Parse TypoLink format into structured object.
- *
- * CRITICAL: Order is url, target, class, title, additionalParams
- * This matches TYPO3's TypoLinkCodecService.php
- *
- * @param {string} typoLink - The TypoLink string
- * @return {Object} Parsed link data with href, target, class, title, additionalParams
- */
-function parseTypoLink(typoLink) {
-    const result = {
-        href: '',
-        target: '',
-        class: '',
-        title: '',
-        additionalParams: ''
-    };
-
-    if (!typoLink || typeof typoLink !== 'string') {
-        return result;
-    }
-
-    typoLink = typoLink.trim();
-    if (typoLink === '') {
-        return result;
-    }
-
-    // Parse using CSV-like logic (space delimiter, quote enclosure, backslash escape)
-    const parts = parseTypoLinkParts(typoLink);
-
-    // Order: url, target, class, title, additionalParams
-    if (parts.length > 0 && parts[0] !== '-') {
-        result.href = parts[0];
-    }
-    if (parts.length > 1 && parts[1] !== '-') {
-        result.target = parts[1];
-    }
-    if (parts.length > 2 && parts[2] !== '-') {
-        result.class = parts[2];
-    }
-    if (parts.length > 3 && parts[3] !== '-') {
-        result.title = parts[3];
-    }
-    if (parts.length > 4 && parts[4] !== '-') {
-        result.additionalParams = parts[4];
-    }
-
-    return result;
-}
-
-/**
- * Encode link data into TypoLink format.
- * This is the reverse of parseTypoLink.
- *
- * Format: url target class title additionalParams
- * - Empty values use '-' placeholder
- * - Values with spaces are quoted
- *
- * @param {Object} linkData - Object with href, target, class, title, additionalParams
- * @return {string} TypoLink string
- */
-function encodeTypoLink(linkData) {
-    const url = linkData.href || '';
-    const target = linkData.target || '-';
-    const cssClass = linkData.class || '-';
-    const title = linkData.title || '-';
-    const additionalParams = linkData.additionalParams || '-';
-
-    // If URL is empty, return empty string
-    if (!url) {
-        return '';
-    }
-
-    // Quote values that contain spaces or special characters
-    const quoteIfNeeded = function(value) {
-        if (value === '-') {
-            return '-';
-        }
-        // Quote if contains space, quote, or backslash
-        if (value.indexOf(' ') !== -1 || value.indexOf('"') !== -1 || value.indexOf('\\') !== -1) {
-            // Escape backslashes and quotes
-            const escaped = value.replace(/\\/g, '\\\\').replace(/"/g, '\\"');
-            return '"' + escaped + '"';
-        }
-        return value;
-    };
-
-    // Build TypoLink parts array
-    const parts = [url];
-
-    // Only include parts up to the last non-empty value
-    // This keeps the output minimal while still correct
-    const hasAdditionalParams = additionalParams !== '-';
-    const hasTitle = title !== '-' || hasAdditionalParams;
-    const hasClass = cssClass !== '-' || hasTitle;
-    const hasTarget = target !== '-' || hasClass;
-
-    if (hasTarget) {
-        parts.push(quoteIfNeeded(target));
-    }
-    if (hasClass) {
-        parts.push(quoteIfNeeded(cssClass));
-    }
-    if (hasTitle) {
-        parts.push(quoteIfNeeded(title));
-    }
-    if (hasAdditionalParams) {
-        parts.push(quoteIfNeeded(additionalParams));
-    }
-
-    return parts.join(' ');
-}
+import {
+    parseTypoLink,
+    parseTypoLinkParts,
+    encodeTypoLink
+} from '../../../Resources/Public/JavaScript/Plugins/typolink-parser.js';
 
 describe('TypoLink Parser', () => {
